Add 3DS support.
3DS payment validation takes an extra step for validation. The payment
starts in the API and Mangopay, but instead of a success or failure
Mangopay returns a URL to redirect the user to a bank webpage to
validate the user preagreed 3DS password on than bank.

Once the user successfully authenticates via 3DS the page will proceed
to the API endpoints for completing the step.

This affects account recharges and subscription payments, so a redirect is
added, along with a completion step endpoint for each of them.

Remember clients MUST normally not call the completion endpoints,
for they will be auto-redirected to after a successful 3DS validation.